+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to [Earthly](https://github.com/earthly/earthly) will be doc
 
 ## Unreleased
 
+### Added
+
+- New `EARTHLY_CI` builtin arg, which is set to `true` when earthly is run with the `--ci` flag. [#2398](https://github.com/earthly/earthly/pull/2398)
+
 ### Changed
 
 - Updated buildkit to include changes up to [a5263dd0f990a3fe17b67e0002b76bfd1f5b433d](https://github.com/moby/buildkit/commit/a5263dd0f990a3fe17b67e0002b76bfd1f5b433d)